Update to 0.3.2
Pardon the errors in version numbers for the last couple of commits.
This update brings slight changes to the main layout, including making
text not as bold, and the navigation bar slightly thinner. I have also
added 5 new pages that now allow for part of the footer to be
functional, including the about us, for parents, for teachers, privacy,
and tos pages.